moved main unrelated methods to Train
