+++ lab2/src/train/Train.java
@@ -5,6 +5,7 @@ import java.text.SimpleDateFormat;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeFormatterBuilder;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Scanner;
@@ -88,6 +89,47 @@ public class Train {
         train.setRailcarNumber(RailcarType.LUXURY, scanner.nextInt());
         return train;
     }
+
+    public static Train[] haveDestination(Train[] trains, String dest){
+        if(trains == null){
+            throw new NullPointerException();
+        }
+        ArrayList<Train> haveDest = new ArrayList<>();
+        for (int i = 0; i < trains.length; i++) {
+            if(trains[i].getDestination().equals(dest)){
+                haveDest.add(trains[i]);
+            }
+        }
+        Train[] result = new Train[haveDest.size()];
+        return haveDest.toArray(result);
+    }
+    public static Train[] haveDestinationAndAfterDateTime(Train[] trains, String dest, LocalDateTime time){
+//        ArrayList<Train> arrayList = new ArrayList<>(Arrays.stream(haveDestination(trains, dest)).toList());
+        Train[] trainsWithDestination = Train.haveDestination(trains, dest);
+        ArrayList<Train> arrayList = new ArrayList<>();
+        for (Train tr:
+                trainsWithDestination) {
+            if(tr.getDepartureTime().isAfter(time)){
+                arrayList.add(tr);
+            }
+        }
+        Train[] result = new Train[arrayList.size()];
+        return arrayList.toArray(result);
+    }
+    public static Train[] haveDestinationAndHaveGeneralCars(Train[] trains, String dest){
+        Train[] trainsWithDestination = Train.haveDestination(trains, dest);
+        ArrayList<Train> arrayList = new ArrayList<>();
+        for (Train tr:
+                trainsWithDestination) {
+            if(tr.getRailcarsNumber(RailcarType.GENERAL) > 0){
+                arrayList.add(tr);
+            }
+        }
+
+        Train[] result = new Train[arrayList.size()];
+        return arrayList.toArray(result);
+    }
+
     @Override
     public String toString() {
         StringBuilder stringBuilder = new StringBuilder();
